+++ gitlint/rules.py
@@ -286,3 +286,23 @@ class BodyChangedFileMention(CommitRule):
                     violation_message = u"Body does not mention changed file '{0}'".format(needs_mentioned_file)
                     violations.append(RuleViolation(self.id, violation_message, None, len(commit.message.body) + 1))
         return violations if violations else None
+
+
+class AuthorValidEmail(CommitRule):
+    name = "author-valid-email"
+    id = "M1"
+    options_spec = [StrOption('regex', "[^@ ]+@[^@ ]+\.[^@ ]+", "Regex that author email address should match")]
+    email_regex = None
+
+    # Note that unicode is allowed in email addresses
+    # See http://stackoverflow.com/questions/3844431/are-email-addresses-allowed-to-contain-non-alphanumeric-characters
+
+    def validate(self, commit):
+
+        # Only compile the regex the first time so that we don't need to re-compile it if we are linting
+        # multiple commits at once
+        if not self.email_regex:
+            self.email_regex = re.compile(self.options['regex'].value, re.UNICODE)
+
+        if commit.author_email and not self.email_regex.match(commit.author_email):
+            return [RuleViolation(self.id, "Author email for commit is invalid", commit.author_email)]